Move DefaultPathResolver/IPathResolver out of VirtualFileSystem and make them a public class.
